test peter

+++ renew_quickstart.sh
@@ -195,7 +195,9 @@ show_menu(){
 		echo " 1 ] Up the Concourse stack"
 		echo " 2 ] Run the Concourse stack by Daemon"
 		echo " 3 ] Down the Concourse stack"
-		echo " W ] Refresh worker pubkey & restart Concourse stack by Daemon"
+		if [[ $(cat svrty.log) == "master" ]]; then
+			echo " W ] Refresh worker pubkey & restart Concourse stack by Daemon"
+		fi
 		echo -e "\033[3;31mRM ] Clear the setup\033[0m"
 		echo -e "$BAR2"
 		echo -n "Please select run type as above: [1-3 or RM] "
@@ -204,9 +206,14 @@ show_menu(){
 			1) docker-compose up ;;
 			2) docker-compose up -d ;;
 			3) docker-compose down ;;
-			W|w) worker_key_consul 
-			     docker-compose down
-			     docker-compose up -d ;;
+			W|w) 
+			if [[ $(cat svrty.log) == "master" ]]; then
+				worker_key_consul 
+				docker-compose down
+				docker-compose up -d ;;
+			else
+				echo  -e "\033[1;31m >> Select Menu\033[0m" ;;
+			fi
 			RM | rm) clear_setup ;;
 			*) echo  -e "\033[1;31m >> Select Menu\033[0m" ;;
 		esac
